# src/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "REST API minta backend",
        "version": "1.0.0",
        "description": "<img alt='DB_diagram' height='300px' src='https://nitslaszlo.github.io/static/session/db_diagram.jpg' />"
    },
    "servers": [
        {
            "url": "http://localhost:5000"
        },
        {
            "url": "https://jedlik.cyclic.app"
        }
    ],
    "paths": {
...
        "/posts": {
            "get": {
                "tags": ["posts"],
                "summary": "Összes post lekérdezése.",
                "description": "Az összes post adatát tudjuk lekérdezni.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/post"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": ["posts"],
                "summary": "Új post mentése.",
                "description": "Új post mentése a 'kategoriak' táblában",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "$ref": "#/components/schemas/recipe"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK (Az új, elmentett kategóriával tér vissza.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/recipe"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error.message"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "post": {
                "type": "object",
                "required": ["title"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A post azonosítója (ObjectID).",
                        "example": "61b5f28df39e4edcf5b8a3e7"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A post szerzője (ObjectID).",
                        "example": "61b5e9c0f39e4edcf5b8a3b9"
                    },
                    "content": {
                        "type": "string",
                        "description": "A post tartalma.",
                        "example": "Frankie Dunn pályája során számos fantasztikus bunyóst edzett és menedzselt. Legfőbb tanítása, mely vezérli az életét is, hogy a legfontosabb az, hogy mindig védd magad. Mióta elidegenedett a lányától, Frankie jó ideje senkit sem engedett közel magához. Egyetlen barátja Hepaj, az egykori bokszoló, aki rendben tartja Frankie edzőtermét, és tudja, hogy a mogorva külső olyan embert takar, aki az elmúlt 23 évben szinte minden nap misére járt, keresve a megbocsátást, amely valahogy mégsem adatik meg neki. Aztán egy nap Maggie Fitzgerald besétál az edzőterembe. Maggie-nek sosem jutott sok semmiből, de van benne valami, ami csak keveseknek adatik meg: tudja, mit akar, és bármit hajlandó megtenni érte. Az élete állandó küzdelem, amit nyers tehetségével, kizökkenthetetlen koncentrálásával és elképesztő akaraterejével képes túlélni."
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 50,
                        "description": "A post címe címe.",
                        "example": "Millió dolláros bébi 2004"
                    }
                }
            },
            "recipe": {
                "type": "object",
                "required": ["recipeName"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "PK - A recept azonosítója (ObjectID).",
                        "example": "61daf48ba5775848e99f555b"
                    },
                    "user_id": {
                        "type": "string",
                        "description": "FK - A recept szerzője (ObjectID).",
                        "example": "61b5e9c0f39e4edcf5b8a3b9"
                    },
                    "recipeName": {
                        "type": "string",
                        "maxLength": 50,
                        "description": "A recept neve.",
                        "example": "PENNE ALL ARRABBIATA CSIRKEMELLEL"
                    },
                    "imageURL": {
                        "type": "string",
                        "description": "Az elkészült ételt ábrázoló kép URL címe.",
                        "example": "https://www.mindmegette.hu/images/240/O/penne-arrabbiata-csirkes.jpg"
                    },
                    "decription": {
                        "type": "string",
                        "description": "Az étel elkészítésének leírása.",
                        "example": "A tésztát a csomágolásán látható utasítás szerint forró, sós vízbe kifőzzük és leszűrjük. A finomra vágott hagymákat olajon megdinszteljük, majd hozzáadjuk a kockára vágott csirkemellet és fehéredésig pirítjuk. Sózzuk, borsozzuk, ráöntjük a paradicsomkonzervet és a vörösbort majd forrásig melegítjük.\nHozzáadjuk a chilit, a felaprított bazsalikomot, az oregánót, végül pedig a barnacukrot és 25-30 percig főzzük, míg a csirkemell meg nem puhul, majd összekeverjük a tésztával."
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "A hozzávalókat tartalmazó tömb.",
                        "example": "[\"50 dkg penne\", \"1 db vöröshagyma\", \"4 gerezd fokhagyma\", \"30 dkg\tcsirkemell\", \"só\", \"bors\", \"stb.\"]"
                    }
                }
            },
...
        }
    }
}